refactor: convert imports to new location for generated types and hooks
