+++ src/pages/groups/[name]/members.tsx
@@ -10,7 +10,7 @@ import { truncate } from "lodash";
 import { NextPage } from "next";
 import { useRouter } from "next/router";
 import { useTranslation } from "react-i18next";
-import { useGroupMembersQuery } from "../../../apollo/gen";
+import { useGroupMembersQuery } from "../../../apollo/groups/queries/gen/GroupMembers.query";
 import GroupMember from "../../../components/Groups/GroupMember";
 import Breadcrumbs from "../../../components/Shared/Breadcrumbs";
 import ProgressBar from "../../../components/Shared/ProgressBar";
