+++ src/pages/roles/index.tsx
@@ -1,12 +1,12 @@
 import { Typography } from "@mui/material";
 import { NextPage } from "next";
 import { useTranslation } from "react-i18next";
-import { useServerRolesQuery } from "../../apollo/gen";
 import ServerRoleForm from "../../components/Roles/ServerRoles/ServerRoleForm";
 import RoleList from "../../components/Roles/RoleList";
 import LevelOneHeading from "../../components/Shared/LevelOneHeading";
 import ProgressBar from "../../components/Shared/ProgressBar";
 import { isDeniedAccess } from "../../utils/error.utils";
+import { useServerRolesQuery } from "../../apollo/roles/queries/gen/ServerRoles.query";
 
 const ServerRoles: NextPage = () => {
   const { data, loading, error } = useServerRolesQuery();
